pyproject.toml
[project]
name = "purrgress"
version = "1.0.0"
description = "personal productivity automation suite 😼"
requires-python = ">=3.9"
dependencies = [
  "click>=8.1",
  "PyYAML>=6.0",
  "pytz>=2024.1",
  "rich>=13.0",
  "questionary>=2.0",
  "pandas>=2.0",
  "matplotlib>=3.8"
]

[project.optional-dependencies]
dev = ["pytest>=8.2", "pytest-mock>=3.14"]

[project.scripts]
purg = "purrgress.cli:cli"
plog = "purrgress.plog_cli:cli"

[build-system]
requires      = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where    = [""]       
include  = ["purrgress*"] 
exclude  = ["tests*", "docs"]